// src/models/baseCaptacionModel.ts
export interface BaseCaptacionType {
    curp: string;  // char(18)
    nombre: string;  // varchar(100)
    primerApellido: string;  // varchar(25)
    segundoApellido: string;  // varchar(25)
    lugarNacimiento: string;  // varchar(100)
    genero: string;  // char(1)
    indigena: string;  // char(1) ('N' en el select)
    unidadReal: string;  // varchar(25)
    unidadOficial: string;  // char(2)
    carrera: string;  // char(5)
    periodo: string;  // char(5)
    discapacidad: string;  // 'N' (hardcoded)
    estatus: 'REGISTRADO SIN VALIDAR' | 'REGISTRADO VALIDADO' | 'EXAMEN PAGADO' | 'PRESENTO EXAMEN' | 'INSCRIPCION PAGADA' | 'INTEGRACION DE EXPEDIENTE' | 'EXPEDIENTE VALIDADO' | 'INSCRITO';  // enum
    estudios: string;  // varchar(100)
    modalidad: string;  // varchar(33)
    municipio: string;  // varchar(100)
    estado: string;  // varchar(100)
    pagoExamen: number;  // tinyint(4)
    pagoInscripcion: number;  // tinyint(4)
    docsEntregados: number;  // tinyint(4)
    inscripcionCompleta: number;  // tinyint(4)
    medioCaptacion: string;  // varchar(100)
}